+++ tests/Model/Domain/DeleteResponse.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * Copyright (C) 2013 Mailgun
+ *
+ * This software may be modified and distributed under the terms
+ * of the MIT license. See the LICENSE file for details.
+ */
+
+namespace Mailgun\Tests\Model\Domain;
+
+use Mailgun\Model\Domain\DeleteResponse as DeleteResponseAlias;
+use Mailgun\Tests\Model\BaseModel;
+
+class DeleteResponse extends BaseModel
+{
+    public function testCreate()
+    {
+        $json =
+        <<<'JSON'
+{
+  "message": "Domain has been deleted"
+}
+JSON;
+        $model = DeleteResponseAlias::create(json_decode($json, true));
+        $this->assertNotEmpty($model->getMessage());
+    }
+}